requirements.txt
anyio==4.9.0
backports-datetime-fromisoformat==2.0.3
beautifulsoup4==4.13.4
certifi==2025.4.26
exceptiongroup==1.2.2
filetype==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Js2Py_3.13==0.74.1
lxml==5.4.0
m3u8==6.0.0
pyjsparser==2.7.1
pyotp==2.9.0
python-dotenv==1.1.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
twikit==2.3.3
typing_extensions==4.13.2
tzlocal==5.3.1
webvtt-py==0.5.1
